Minor Adjustments
Added sign up and register text.
Made the first  loading page to be landing.
