+++ Stix/Stix/Pages/Index.cshtml.cs
@@ -1,36 +1,10 @@
-﻿using System.Globalization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace Stix.Pages
 {
-    public class IndexModel : PageModel
+    public class LandingModel : PageModel
     {
-        private readonly ILogger<IndexModel> _logger;
-
-        public IndexModel(ILogger<IndexModel> logger)
-        {
-            _logger = logger;
-        }
-        
-        [BindProperty]
-        public string incidentName { get; set; } = "Default name";
-        [BindProperty]
-        public string description { get; set; } = "This section describes the incident";
-        [BindProperty]
-        public string addedOn { get; set; } = "2022-02-22\n" +
-            "T69:42:00";
-        [BindProperty]
-        public string viewMore { get; set; } = "";
-
-        public void OnGet()
-        {
-            
-        }
-
-        public void OnPost() 
-        { 
-            
-        }
+       
     }
 }
