--- Stix/Stix/Pages/Landing.cshtml
@@ -1,44 +1,0 @@
-﻿@page
-@model Stix.Pages.LandingModel
-@{
-  Layout = null;  
-}
-
-<html lang="en">
-<head>
- <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-      <title>Page Title</title>
-<link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
-    <link rel="stylesheet" href="~/Stix.styles.css" asp-append-version="true" />
-        <link rel="stylesheet" href="~/css/Landing.css">
-</head>
-<body>
-    <div class="container-fluid">
-        <div class="row">
-            <div class="col-sm-8"></div>
-            <div class="col-sm-4 my-custom-col align-items-center">
-                <div class="row-sm-20 push">
-
-                </div>
-                <div class="row-sm-20 content">
-                   
-                <h1 id="stixHeader">STIX Database</h1>
-                 <p id="stixParagraph">The place for maritime cyber<br>
-                  incidents reporting</p>
-                  
-                  <a href="Login">  
-                  <button id="logRegButt"> Login/Register  </button>
-                  </a>
-                  </div>
-             </div>
-        </div>
-
-    </div>
-<script src="~/lib/jquery/dist/jquery.min.js"></script>
-    <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-    <script src="~/js/site.js" asp-append-version="true"></script>
-
-</body>
-</html>
