--- Stix/Stix/Pages/Landing.cshtml.cs
@@ -1,10 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace Stix.Pages
-{
-    public class LandingModel : PageModel
-    {
-       
-    }
-}
