+++ Stix/Stix/Pages/Overview.cshtml
@@ -1,0 +1,31 @@
+﻿@page
+@model IndexModel
+@{
+    ViewData["Title"] = "STIX";
+}
+
+    <div class="filterSortContainer">
+        <a href=""><button type="button">Sort</button></a>
+        <a href=""><button type="button">Filter</button></a>
+    </div>
+    
+    <div class="overviewTableContainer flex">
+        <table class="overviewTable">
+            <tr>
+              <th class="incidentHeader">Incident Name</th>
+              <th class="descriptionHeader">Description</th>
+              <th class="addedOnHeader">Added On</th>
+              <th class="viewMoreHeader">View More</th>
+            </tr>
+        <!-- loop, can change it to just grab everything from a db-->
+        @for (int i = 0; i < 15 ; i++)
+        {
+           <tr>
+              <td class="leftTableCell">@Model.incidentName</td>
+              <td class="centerTableCell">@Model.description</td>
+              <td class="centerTableCell">@Model.addedOn</td>
+              <td class="rightTableCell">"x"</td>
+            </tr> 
+        }
+          </table>
+    </div>
