+++ Stix/Stix/Pages/RegisterModerator.cshtml
@@ -50,7 +50,12 @@
         <input type="text-area" asp-for="Moderator.reason" class="form-control" id="registerReasonModerator"
             placeholder="Enter a reason/proof(field of expertise, degrees, etc.)">
     </div>
-
+    <a href="register"> 
+    <h6> Register... </h6> 
+    </a>
+    <a href="login"> 
+    <h6> Login.. </h6> 
+    </a>
     <button type="submit" class="btn btn-primary">Submit</button>
 </form>
 </div>
